backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-assistant-backend"
version = "0.1.0"
description = "Financial Assistant with OCR and Expense Management"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0",
    "langchain>=0.0.300",
    "langgraph>=0.0.10",
    "supabase>=1.0.4",
    "google-generativeai>=0.3.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0"
    ,"bcrypt>=4.0.1"
    ,"PyJWT>=2.8.0"
    ,"python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.6",
    "black>=23.0.0",
    "mypy>=1.5.0"
]

[tool.ruff]
line-length = 88
extend-ignore = [
    "E501",  # Line too long (handled by black)
]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
extend-per-file-ignores = {}
# Exclude a variety of commonly ignored directories.
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 88
target-version = ['py310']
skip-string-normalization = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true